% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_adjacent.R
\name{decision_tree_adj}
\alias{decision_tree_adj}
\title{Make decisions for potential duplicates identified by \code{\link[=dup_find_fuzzy_adj]{dup_find_fuzzy_adj()}}}
\usage{
decision_tree_adj(df, id_dup_pair)
}
\arguments{
\item{df}{A data frame (i.e., output #1 of \code{\link[=dup_find_fuzzy_adj]{dup_find_fuzzy_adj()}}).}

\item{id_dup_pair}{A data frame listing \code{id} of potential duplicate pairs (i.e., output #2 of \code{\link[=dup_find_fuzzy_adj]{dup_find_fuzzy_adj()}})}
}
\value{
Two data frames. (1) \code{df} with \code{"match"} column modified according to the decision tree; (2) \code{id_dup_pair} with \code{"decision"} column added.
}
\description{
Make decisions for potential duplicates identified by \code{\link[=dup_find_fuzzy_adj]{dup_find_fuzzy_adj()}}
}
\examples{
\dontrun{
c(df, id_dup_pair) \%<-\% decision_tree_adj(df, id_dup_pair)
}
}
