% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_pairwise.R
\name{decision_tree_pairwise}
\alias{decision_tree_pairwise}
\title{Make decisions for potential duplicates}
\usage{
decision_tree_pairwise(ls_df, id_dup_pair)
}
\arguments{
\item{ls_df}{A list of data frames containing the partitioned dataset  (i.e., output #1 of \code{\link[=simi_ptn_pair]{simi_ptn_pair()}}).}

\item{id_dup_pair}{A data frame listing record id and partition id of duplicate pairs (i.e., output of \code{\link[=dup_find_fuzzy_pairwise]{dup_find_fuzzy_pairwise()}}).}
}
\value{
\code{id_dup_pair} with \code{"decision"} column added.
}
\description{
Make decisions for potential duplicates
}
\examples{
\dontrun{
id_dup_pair <- decision_tree_pairwise(ls_df, id_dup_pair)
}
}
