% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_match.R
\name{dedu_exact}
\alias{dedu_exact}
\title{Find duplicates by exact match and remove them}
\usage{
dedu_exact(df, match_by)
}
\arguments{
\item{df}{A data frame with bibliographic information that has gone through text normalization. \code{df} must have the following 6 columns \code{c("author", "title", "journal", "abstract", "year", "doi")}.}

\item{match_by}{Quoted name(s) of the column(s) with information to match by (e.g., \code{"doi_norm"}, \code{"title"}, \code{"title_norm"}, \code{c("doi_norm", "title", "title_norm")}).

If supplying a character vector with \emph{multiple} elements, deduplication will be performed in order. For example, if \code{match_by = c("doi_norm", "title", "title_norm")}, deduplication will be performed first according to \code{"doi_norm"}, then \code{"title"}, and finally \code{"title_norm"}.}
}
\value{
Deduplicated \code{df}.
}
\description{
\itemize{
\item This automatically removes duplicates identified by exact match without manual review.
\item Support deduplication based on multiple columns (one at a time).
}
}
\details{
Records with missing information (i.e., NA) in \code{match_by} column won't be modified.
}
\examples{
# load example dataset
data(bib_example_small)

# text normalization of the data frame
df <- norm_df(bib_example_small)

# deduplicate according to 3 columns in order (one at a time)
df_new <- dedu_exact(df, match_by = c("doi_norm", "title", "title_norm"))
}
