% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_match.R
\name{dup_find_exact}
\alias{dup_find_exact}
\title{Find duplicates by exact match}
\usage{
dup_find_exact(
  df,
  match_by,
  double_check = FALSE,
  check_by = "first_author_last_name_norm"
)
}
\arguments{
\item{df}{A data frame with bibliographic information that has gone through text normalization. \code{df} must have the following 6 columns \code{c("author", "title", "journal", "abstract", "year", "doi")}.}

\item{match_by}{Quoted name of the column with information to match by (e.g., \code{"doi_norm"}, \code{"title"}, \code{"title_norm"}).}

\item{double_check}{Logical: Is confirmarion against another column needed? Defaults to \code{FALSE}.}

\item{check_by}{Quoted name of the column with information to double check by. Not required/ignored if \code{double_check == FALSE}. Defaults to \code{"first_author_last_name_norm"} if \code{double_check == TRUE}.}
}
\value{
\itemize{
\item If \code{double_check == FALSE}, return \code{df} with a new column named \code{"match"}.
\item If \code{double_check == TRUE}, return 2 data frames (\code{df} and \code{df_manual_check}). Syntax \verb{\%<-\%} must be used in this case to have the function return 2 data frames.
}
}
\description{
Find duplicates by exact match
}
\details{
\itemize{
\item Records with missing information (i.e., NA) in \code{match_by} column won't be modified.
\item \strong{Double check criteria:}
Within each identified duplicate set, if \code{check_by} is the same, remain as duplicates without review; if \code{check_by} is different, output the duplicate set in \code{df_manual_check} for manual review.
}
}
\examples{
library(zeallot)  # to use `\%<-\%`

# load example dataset
data(bib_example_small)

# text normalization of the data frame
df <- norm_df(bib_example_small)

# example 1:
df_1 <- dup_find_exact(df, match_by = "doi_norm", double_check = FALSE)
# Alternatively, \%<-\% will also work
df_2 \%<-\% dup_find_exact(df, match_by = "doi_norm", double_check = FALSE)

# example 2:
c(df, df_manual_check) \%<-\% dup_find_exact(df, match_by = "title_norm", double_check = TRUE, check_by = "first_author_last_name_norm")  # Syntax \%<-\% must be used in this case to have the function return 2 data frames.
}
