% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_find_pairwise.R
\name{dup_find_fuzzy_pairwise}
\alias{dup_find_fuzzy_pairwise}
\title{Find duplicates by fuzzy match of string similarity between pairwise records}
\usage{
dup_find_fuzzy_pairwise(
  ls_df,
  ls_df_simi,
  cutoff_title = 0.7,
  cutoff_abstract = 0.7
)
}
\arguments{
\item{ls_df}{A list of data frames containing the partitioned dataset  (i.e., output #1 of \code{\link[=simi_ptn_pair]{simi_ptn_pair()}}).}

\item{ls_df_simi}{A list of data frames with string similarity results calculated (i.e., output #2 of \code{\link[=simi_ptn_pair]{simi_ptn_pair()}}).}

\item{cutoff_title}{Cutoff threshold of string similarity for normalized title. Range: [0, 1]. Defaults to 0.7.}

\item{cutoff_abstract}{Cutoff threshold of string similarity for normalized abstract. Range: [0, 1]. Defaults to 0.7.}
}
\value{
A data frame listing record id and partition id of duplicate pairs.
}
\description{
Find duplicates by fuzzy match of string similarity between pairwise records
}
\details{
\strong{For both cutoffs:}

Cutoff thresholds in \code{\link[=dup_find_fuzzy_adj]{dup_find_fuzzy_adj()}} are usually applicable here. Alternatively, you can re-examine the similarity distribution plots by \code{\link[=plot_simi_dist]{plot_simi_dist()}} and choose sensible values.
}
\examples{
\dontrun{
id_dup_pair <- dup_find_fuzzy_pairwise(ls_df, ls_df_simi, cutoff_title = 0.7, cutoff_abstract = 0.7)
}
}
