% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_resolve_pairwise.R
\name{dup_resolve_pairwise}
\alias{dup_resolve_pairwise}
\title{Manually resolve potential duplicate pairs requiring "check"}
\usage{
dup_resolve_pairwise(id_dup_pair, df_check, match_index, result)
}
\arguments{
\item{id_dup_pair}{A data frame listing record id and partition id of duplicate pairs after getting automatic decisions (i.e., output of \code{\link[=decision_tree_pairwise]{decision_tree_pairwise()}}).}

\item{df_check}{A data frame of duplicate pairs for manual review (i.e., output of \code{\link[=dup_screen_pairwise]{dup_screen_pairwise()}}).}

\item{match_index}{Numeric: a vector of \code{"match"} numbers to change the decision for.}

\item{result}{Character: "duplicate" or "not duplicate".}
}
\value{
A data frame: the input \code{id_dup_pair} with \code{"decision"} column modified accordingly.
}
\description{
Change decision to "duplicate" or "not duplicate" according to manual review results.
}
\examples{
\dontrun{
id_dup_pair <- dup_resolve_pairwise(
id_dup_pair,
df_check,
match_index = c(1, 2, 3, 4),
result = "not duplicate")
}
}
