% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_rm_adjacent.R
\name{dup_rm_adj}
\alias{dup_rm_adj}
\title{Remove duplicates between adjacent rows}
\usage{
dup_rm_adj(df, id_dup_pair)
}
\arguments{
\item{df}{A data frame (i.e., output #1 of \code{\link[=decision_tree_adj]{decision_tree_adj()}})}

\item{id_dup_pair}{A data frame listing \code{id} of potential duplicate pairs (i.e., output #2 of \code{\link[=decision_tree_adj]{decision_tree_adj()}})}
}
\value{
The input \code{df} with duplicates removed.
}
\description{
This function ensures that the most recent record be kept. When a peer-reviewed publication co-exists with a preprint or a conference proceeding, the peer-reviewed version will be kept.
}
\examples{
\dontrun{
df_2 <- dup_rm_adj(df, id_dup_pair)
}
}
