% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_rm_pairwise.R
\name{dup_rm_pairwise}
\alias{dup_rm_pairwise}
\title{Remove duplicates in pairwise comparison}
\usage{
dup_rm_pairwise(ls_df, id_dup_pair, to_dataframe = TRUE)
}
\arguments{
\item{ls_df}{A list of data frames containing the partitioned dataset  (i.e., output #1 of \code{\link[=simi_ptn_pair]{simi_ptn_pair()}}).}

\item{id_dup_pair}{A data frame listing record id and partition id of duplicate pairs after resolving checked duplicates (i.e., output of \code{\link[=dup_resolve_pairwise]{dup_resolve_pairwise()}}).}

\item{to_dataframe}{Logical: Should we merge the list of data frames into a single data frame? Defaults to \code{TRUE}.}
}
\value{
The input \code{ls_df} with duplicates removed. The resulted list of data frames are merged into a single data frame if \code{to_dataframe == TRUE}. Otherwise, a list of data frames is returned.
}
\description{
This function ensures that the most recent record be kept. When a peer-reviewed publication co-exists with a preprint or a conference proceeding, the peer-reviewed version will be kept.
}
\examples{
\dontrun{
df_2 <- dup_rm_pairwise(ls_df, id_dup_pair, to_dataframe = TRUE)
# or
ls_df_2 <- dup_rm_pairwise(ls_df, id_dup_pair, to_dataframe = FALSE)
}
}
