% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_helper_field.R
\name{extract_initialism}
\alias{extract_initialism}
\title{Extract initialism
Extract 1st letter of each word and delete all the other letters}
\usage{
extract_initialism(string)
}
\arguments{
\item{string}{A character vector (e.g., a column in a data frame)}
}
\value{
A character vector
}
\description{
Extract initialism
Extract 1st letter of each word and delete all the other letters
}
\examples{
# Example 1
journal <- c("JOURNAL OF BIOLOGICAL CHEMISTRY",
             "Proteins: Structure, Function and Bioinformatics",
             "Archives of Biochemistry and Biophysics",
             "Clin Infect Dis",
             "Clinical Infectious Diseases",
             "CLINICAL INFECTIOUS DISEASES")

extract_initialism(journal)


# Example 2
journal_norm <- c("journal biological chemistry",
                  "proteins structure function bioinformatics",
                  "archives biochemistry biophysics",
                  "clin infect dis",
                  "clinical infectious diseases",
                  "clinical infectious diseases")

extract_initialism(journal_norm)
}
