% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_normalization.R
\name{norm_author}
\alias{norm_author}
\title{Clean and normalize author in bibliography}
\usage{
norm_author(author, rm_punctuation = FALSE)
}
\arguments{
\item{author}{A character vector (e.g., a column in a data frame)}

\item{rm_punctuation}{Logical: Does unaccenting characters introduce extra punctuation? If so, these need to be removed. Defaults to \code{FALSE}.

Using \verb{bash iconv} to unaccent characters will introduce extra punctuation (e.g., '`^~\"). If using bash iconv, these punctuation needs to be removed as well. Since we use python now, this is not needed by default.}
}
\value{
Normalized character vector
}
\description{
For author, we do the following string normalization.
\itemize{
\item replace “and” with “&” (Reduce the effect on dissimilarity)
\item remove all space (1. Remove extra whitespace  2. Reduce the effect of space-caused dissimilarity)
\item convert letters to lowercase
}
}
\examples{
# Example 1
author <- c("Xia, Z. X. and Dai, W. W. and Xiong, J. P. and Hao, Z. P. and Davidson, V. L. and White, S. and Mathews, F. S.",
"Ahmed, M. H. and Koparde, V. N. and Safo, M. K. and Neel Scarsdale, J. and Kellogg, G. E.",
"Whitman, C. P." )

norm_author(author)


# Example 2
# é becomes 'e if you use `cat file.bib | iconv -f utf8 -t ascii//TRANSLIT//IGNORE > convert.bib` to unaccent characters. Make `rm_punctuation = TRUE` to remove the extra punctuation intruduced.
author2 <- c("Ren'ee")

norm_author(author2, rm_punctuation = TRUE)

}
