% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_normalization_df.R
\name{norm_df}
\alias{norm_df}
\title{Clean and normalize the entire data frame with bibliographic information}
\usage{
norm_df(df)
}
\arguments{
\item{df}{A data frame with bibliographic information}
}
\value{
A normalized data frame. New columns containing normalized and extracted data are added to the original data frame. By default, expect 8 more columns compared with the original data frame.
}
\description{
Normally we have bibliographic information imported as a data frame rather than individual character vectors. This function \code{norm_df} wraps all (1) text normalization and (2) helper field extraction that are needed in the downstream reference deduplication.
}
\details{
In detail, it normalizes
\itemize{
\item doi (simply convert to lowercase; add column \code{doi_norm})
\item title (see sub-function \code{\link[=norm_title]{norm_title()}}; add column \code{title_norm})
\item author (see sub-function \code{\link[=norm_author]{norm_author()}}; add column \code{author_norm})
\item journal (see sub-function \code{\link[=norm_journal]{norm_journal()}}; add column \code{journal_norm})
\item year (simply convert data type from character to integer; do not add any column)
\item abstract (see sub-function \code{\link[=norm_abstract]{norm_abstract()}}; add column \code{abstract_norm})
}

Additionally, it extracts
\itemize{
\item \code{first_author_last_name} and \code{first_author_last_name_norm}
\item \code{journal_initialism} (apply function \code{\link[=extract_initialism]{extract_initialism()}} to column \code{journal_norm}).
}
}
\examples{
data(bib_example_small)

df_new <- norm_df(bib_example_small)


}
\seealso{
Its sub-functions:
\itemize{
\item \code{\link[=norm_title]{norm_title()}}
\item \code{\link[=norm_author]{norm_author()}}
\item \code{\link[=norm_journal]{norm_journal()}}
\item \code{\link[=norm_abstract]{norm_abstract()}}
\item \code{\link[=extract_initialism]{extract_initialism()}}
}
}
