% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_normalization.R
\name{norm_journal}
\alias{norm_journal}
\title{Clean and normalize journal in bibliography}
\usage{
norm_journal(journal)
}
\arguments{
\item{journal}{A character vector (e.g., a column in a data frame)}
}
\value{
Normalized character vector
}
\description{
For journal, we do the following string normalization.
\itemize{
\item convert letters to lowercase
\item remove punctuation
\item remove English stop words
\item remove whitespace from start and end of string; also reduce repeated whitespace inside the string.
}
}
\examples{
journal <- c("Proteins: Structure, Function and Bioinformatics",
"Zoonoses Public Health",
"Zoonoses and Public Health")

norm_journal(journal)

}
