% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_normalization_others.R
\name{norm_title}
\alias{norm_title}
\title{Clean and normalize title in bibliography}
\usage{
norm_title(title)
}
\arguments{
\item{title}{A character vector (e.g., a column in a data frame)}
}
\value{
Normalized character vector
}
\description{
For title, we do the following string normalization.
\itemize{
\item remove trademark "(TM)"
\item convert letters to lowercase
\item remove punctuation
\item remove all space
}
}
\examples{
title <- c("Evaluation of the Abbott RealTime (TM) CT assay with the BD ProbeTec (TM) ET assay for the detection of Chlamydia trachomatis in a clinical microbiology laboratory",
"Evaluation of the Abbott RealTime CT assay with the BD ProbeTec ET assay for the detection of Chlamydia trachomatis in a clinical microbiology laboratory",
"Evaluation of the Abbott RealTime(tm) CT assay with the BD ProbeTec(tm) ET assay for the detection of Chlamydia trachomatis in a clinical microbiology laboratory",
"beta-lactam Resistance in Pseudomonas aeruginosa: Current Status, Future Prospects" )

norm_title(title)

}
