% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_normalization_transliteration.R
\name{norm_transliteration}
\alias{norm_transliteration}
\title{Transliterate a text file}
\usage{
norm_transliteration(
  file_input,
  file_output,
  method = c("greek_letter-name", "any-ascii"),
  customized
)
}
\arguments{
\item{file_input}{Directory/Path to the input file.}

\item{file_output}{Directory/Path to the out file. No need to create the file in advance. A file with the specified name will be created.}

\item{method}{A character vector specifying the transliteration rules in order. Defaults to a recommended rule combination - \code{c("greek_letter-name", "any-ascii")}. See \code{\link[stringi:stri_trans_general]{stringi::stri_trans_general()}} documentation for more rules in the database. Customized rules are also allowed (see examples).}

\item{customized}{Logical. A logical vector (with the same length as \code{method}) specifying whether the corresponding transliteration rule is a customized rule or not. \emph{Optional if none of the rules is customized (i.e., Defaults to an all-FALSE vector.).}}
}
\description{
Transliterate a text file and output to another text file.
}
