% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simi_distribution.R
\name{plot_simi_dist}
\alias{plot_simi_dist}
\title{Generate similarity distribution plot}
\usage{
plot_simi_dist(df_simi, simi_param)
}
\arguments{
\item{df_simi}{A data frame with string similarity results.}

\item{simi_param}{A character of the similarity parameter to plot the distribution (i.e., quoted name of the column containing the similarity scores). For example, \code{"title_simi"}, \code{"abstract_simi"}, \code{"author_simi"}, \code{"first_author_last_name_simi"}.}
}
\value{
A scatterplot of distributions of similarity scores
}
\description{
Generate similarity distribution plot
}
\examples{
\dontrun{
# `df_simi` is the data frame resulted from `simi_order_df()`
# see the example in the `simi_order_df()` help page for how `df_simi` is generated

# Distribution of similarity scores based on normalized title
p_ti <- plot_simi_dist(df_simi, "title_simi")
p_ti  # show p_ti in the Plots tab

# Distribution of similarity scores based on normalized abstract
p_ab <- plot_simi_dist(df_simi, "abstract_simi")
p_ab  # show p_ab in the Plots tab
}
}
