% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefDeduR_example_dataset}
\alias{RefDeduR_example_dataset}
\alias{bib_example_complete}
\alias{bib_example_small}
\title{Sample data frames containing bibliographic information}
\format{
Data frames containing bibliographic information with variables such as title, author, journal, volume, abstract. Data type of all columns is character.

\code{bib_example_complete} contains 6384 rows. \code{bib_example_small} is a subset of \code{bib_example_complete} and contains only 12 rows.
}
\source{
\code{bib_example_complete} contains all bibliographic records from searching indoor surface microbiome studies on PubMed, Web of Science, and Scopus on 2022-01-10.
}
\usage{
bib_example_complete

bib_example_small
}
\description{
Data frames containing bibliographic information imported from BibTeX files (file.bib) by \code{\link[revtools:read_bibliograpy]{revtools::read_bibliography()}}
}
\keyword{datasets}
