% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simi_distribution.R
\name{plot_simi_dist}
\alias{plot_simi_dist}
\title{Generate similarity distribution plot}
\usage{
plot_simi_dist(df_simi, simi_param)
}
\arguments{
\item{df_simi}{A data frame with string similarity results.}

\item{simi_param}{A character of the similarity parameter to plot the distribution (i.e., quoted name of the column containing the similarity scores). For example, \code{"title_simi"}, \code{"abstract_simi"}, \code{"author_simi"}, \code{"first_author_last_name_simi"}.}
}
\value{
A scatterplot of distributions of similarity scores
}
\description{
Generate similarity distribution plot
}
